Implemented most of CFixedPoint. Replaced boost::optional with CDatabaseNullable, so all external interfaces expose internal objects. Removed CDatabaseValue::FromString and CDatabaseValuePolicy::FromStr.

+++ Source/Plugin/SQLite/Src/DatabaseStatementParameterSetter.h
@@ -114,7 +114,7 @@ BEGIN_NAMESPACE_DATABASE_SQLITE
 		//!@copydoc SSqliteBindingBase::DoSetValue
 		virtual void UpdateValue()
 		{
-			SQLiteTry( SQLite::BindInt( _statement, _index, _value.GetValue() ), StringStream() << STR( "Parameter set value: " ) << _value.GetValue(), EDatabaseExceptionCodes_StatementError, _connection );
+			SQLiteTry( SQLite::BindInt( _statement, _index, _value.GetValue() ), StringStream() << STR( "Parameter set value: " ) << int16_t( _value.GetValue() ), EDatabaseExceptionCodes_StatementError, _connection );
 		}
 
 		//! The parameter value
